!!!! DON'T MERGE THIS YET. !!!! This is a test I was doing on an easier way to have the authentication and registration work Please zip your current project, download this one and see what you think. I had to delete a lot of your code to work on this, so I don't want to merge this unless you are cool with what I did, and then you can copy some of the code you have already written where it needs to go.
What was done:
Implemented Firebase pre-built Sign in/Registration UI
Reduced number of pages for authentication and sign in to 1 page
Google sign in/registration works
Email sign in/registration works
Facebook - I didn't get to test this because it had to be on a device, but I'm fairly certain it will work
One issue I had that I didn't get a chance to look into resolving was after the first time giving app permission to use Google auth, it shouldn't ask you anymore unless you uninstall the app. Currently it is asking every time.